backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
xmltodict==0.13.0
fake-useragent==1.5.1
openai>=1.52.0
python-multipart==0.0.12
apscheduler==3.10.4
selenium==4.25.0
webdriver-manager==4.0.2
pyperclip==1.9.0
pywin32
pyautogui==0.9.54
Pillow==10.4.0
imagehash==4.3.1
python-dotenv==1.0.1
aiofiles==23.2.1
# ffmpeg must be installed separately and available in PATH (video_maker.py)
# Windows: winget install ffmpeg  or  https://ffmpeg.org/download.html
